Fix:  ArithmeticError: Could not align to closest sector (#561278)
Given a partition layout like this:

partition startsector endsector
/boot     2048        1026047
PV        1026048     xxxxxxxxxx

Then when we call getFreeRegions() parted gives us a free region with a
geometry starting at sector 63 and ending at sector 2047, which we then try to
align to our 2048 grain, which fails for the starting sector. When we fail to
align, this means that the free region is too small to create an aligned
partition in there, and we should just skip it.