Clarify which storage exceptions are bugs (#557928).
In turnOnFilesystems, some exceptions are bugs in anaconda and some are just
fatal errors reported by the underlying tools.  An example of the latter is
not being able to resize a filesystem due to space issues.  These should not
get the exception dialog since that looks like anaconda is at fault.  For all
other cases, it's likely an anaconda bug and should get the exception
dialog.