Improve reboot modes in init.c and shutdown.c. (#564297).
Stop the kernel correctly on HALT (so that we dont see 'kernel panic, they
killed init' on i386 and on s390 one gets easy access to manual IPL).

Introduce a new reboot method that does allow us to see the backtrace and
doesn't scroll the screen up with useless unmount info.