Add udev_settle after setup of LUKSDevice. (#534043)
LVM (when looking for PVs) is finding the temporary udev device
instead of the real/final device, so this is to make sure things
are as they should be when lvm goes-a-hunting for devices.

We should probably be explicitly telling lvm which devices to look
for PVs on to avoid it scanning every block device every time we
run an lvm command.