Use property() so we can assign to anaconda.intf (#565639).
I'd rather add a rescue mode case to the existing intf property, but we need
a screen instance to create the rescue mode interface and we then also need
that instance for other things.  This makes it more difficult to hook up when
I could just use the other style of property instead.