Fix hiding the advanced button on the filter UI (#555769, #565425, #560016).
Unrealizing the button box creates a mystery hidden window that causes two
problems.  First, it steals the focus so keyboard controls don't work on
the filtering UI.  Second, it causes a lot of gtk warnings on tty1.  Instead
of unrealizing, the fix is to just hide and disallow show_all working on
it later.