Add [images-xen] block to .treeinfo (#559347).
Under RHEL-6, we do not have a separate xen kernel package nor do we
have a separate kernel-PAE package.  All installations use the 'kernel'
package.  Introduce a $xen variable to control whether or not we want
the [images-xen] block in the .treeinfo file, which fixes virt-install
runs on previous RHEL releases (those installs are looking for an
[images-xen] block).