Move some kickstart-specific storage init into storageInitialization.
The external storage testing work is using kickstart to specify what partitions
to set up.  In order for kickstart to work correctly, all this code would need
to be duplicated in every test case.  It's also pretty annoying that it's a
part of __main__ instead of storageInitialize.  Therefore, move it.