Redo the 'sshd' flag.
Previously, we would only start sshd if you'd passed --kickstart to
anaconda; but, if passed a kickstart file, we would start sshd unless
you explicitly turned it off.  Instead, default sshd to off for both
slow and kickstart installs, and allow it to be enabled for slow
installs from the kernel command line.

Note that, in the absence of a kickstart file specifying users, this
will start sshd with no password for root.  Treat this like you would
treat 'vnc' with no 'vncpasswd'.