Filter UI do not start / stop BIOS RAID sets to get there size (#574682)
Do not start / stop BIOS RAID sets to get there size, this causes
issues with BIOS RAID 5 sets as these are not supported by device mapper
raid in the kernel. Despite this we should still show them (and certainly not
traceback as we do now) as Intel RAID5 sets will work fine (they use mdraid
except for in the filter UI).