yum requires the proxy settings to include a protocol (#484788, #567982).
Since yum requires the protocol and libcurl doesn't care if there's a protocol
given or not, add the protocol to the front of the proxy value itself.  Also,
both yum and libcurl will process the proxy setting if it's got a port on the
end so remove all the special port handling.