BIOS RAID sets get shown double when adding advanced storage (#574714)
FilterWindow.populate() will scan for BIOS RAID sets and add them on
every invocation. This causes them to be added multiple times if
FilterWindow.populate() get called multiple times. This happens for
example when adding an iSCSI target using the "add Advanced target"
button on a machine which has a BIOS RAID set.

Also since the second populate call will not have any of the raid set
members as argument, fstype cannot be determined, and the set shows
up in the Basic Devices tab rather then in the Firmware RAID tab.