Make python start with correct default unicode encoding (#539904).
Currently it's ascii and that's bad. import gtk; changes it to utf-8
anyway but that doesn't happen in text install of course. The default
encoding is set vie sitecustomize.py, it is not possible to do so in any
other way.

The patch has been tested with a custom compose.