Try to pull in generic libraries as well as optimized ones (#572178)
Attempt to pull in generic libraries whenever we see a platform
optimized one.  This implementation uses the naive approach of just
chopping directories out of the middle of library paths and seeing if
there's a file there without that part.  If there is, it pulls that in
as well.

It seems to work to build power5-able trees on a power6 machine.