Use /sbin/dasd_cio_free to free blacklisted DASDs (#558881)
Use the dasd_cio_free command to free blacklisted DASDs and block until
they are available to set online and set other attributes.  We were just
writing "free $DEVICE" to /proc/cio_ignore, but we lacked the wait loop
to make sure the device showed up.  The /sbin/dasd_cio_free command does
that for us, so we can just call that.

NOTE:  The dasd_cio_free command reads the DASD list from the output of
the 'modprobe --showconfig' command, so the $DASD list is placed in the
/etc/modprobe.d/dasd_mod.conf in module option format so that modprobe
can return it correctly.