Remove dasdSetup() from loader (#558881)
The dasdSetup() function in hardware.c is no longer necessary with the
current linuxrc.s390.  The linuxrc.s390 handles all DASD setup except
dasdfmt.  If a DASD needs dasdfmt, that's handled by the anaconda
storage code.

The dasdSetup() function would take the contents of /tmp/dasd_ports and
pass it as the dasd= parameter during a forced reload of the dasd_mod
kernel module.  We do not need to do this anymore since linuxrc.s390
sets up the DASD devices before loader starts.