Write rd_CCW when root fs is on a network device on s390x (#577193)
The dracutSetupString() method in the Network class needed to write out
the rd_CCW parameter, which is necessary for network devices on s390x.
The format is described in the dracut(8) man page.  The required
information has always been written to the ifcfg-DEVICE file, so we just
need to make sure it gets to dracut as well.