Don't make UEFI boot media for i386 (576580)
We're not supporting UEFI on non-64-bit x86 machines in RHEL-6, so don't
make media that tries to support that.

This patch also conditionalizes the use of UEFI on boot.iso, and turns it
off on 64-bit machines, so that boot.iso can be used as a fallback since
disc1 is UEFI enabled.