Restore xdriver=<driver> functionality (#577312)
We used to write /etc/X11/xorg.conf during the initialization of anaconda and
then another one to /mnt/sysimage/etc/X11/xorg.conf during the configuration of
the installed system.

We changed it and the file is now being written to
/mnt/sysimage/etc/X11/xorg.conf on both calls which is obviously wrong.

So this fix restores the od behaviour and should make xdriver=vesa work again.