Take the request's format into account when deciding to resize (#578471).
If the request has a format, check if the format is resizable before offering
a resize option in the UI.  This prevents us from offering to resize PVs,
which anaconda does not support.

There's a related problem this does not fix:  If you choose to reformat a PV
as a regular filesystem, the option to resize does not appear.