Move 'make bumpver' functionality to scripts/makebumpver
Added a Python program to implement the make bumpver functionality.  The
version number is incremented and the RPM spec file changelog is
constructed.  On RHEL branches, each git commit is checked to see that
it references a RHEL bug and that the RHEL bug is in MODIFIED, has the
Fixed In Version field set correctly, and that the bug is for 'Red Hat
Enterprise Linux'.  The idea here is to catch these bug problems before
getting to dist-cvs.

For non RHEL branches, the RPM log is constructed as is, no bug
verification happens.

In instances where you need to ignore certain git commits, you can pass
the -i options to the script with a comma separated list of git commit
IDs to ignore.  This is intended for cases where we need to change a
BuildRequire in the spec file template or something like that.