Add bug mapping support to scripts/makebumpver.
We occassionally have Fedora BZ numbers reference in RHEL branch
commits.  Once the change is pushed, we can't update the commit log.  We
clone the bug in BZ and then can reference the RHEL bug in the spec file
changelog.  The -m option lets you do that.  Using 'make bumpver'
target, here's an example:

    make bumpver IGNORE=ad32abd1612b747aa3dd5c67e3da390b5c9e0d34 \
                 MAP="572488=580404"

This example also shows the IGNORE functionality.  If there are utility
commits (e.g., to makebumpver), we can ignore those during the spec file
changelog generation.  Just pass the git commit ID(s) you want to ignore
as a comma-separated list.

The MAP functionality is FEDORA_BZ=RHEL_BZ and is for instances where we
have a commit on a RHEL branch that references a Fedora bug.  We clone
the bug to make a RHEL bug, then use MAP= to make sure the RHEL bug ends
up in the spec file changelog.