Add a test case to verify that kickstart commands use the right handler.
The aim of this is to prevent bugs like 581829, where a later version of a
command is added in pykickstart but anaconda is never updated to use the
right version.  Of course, this shifts the burden from updating kickstart.py
to running the test case.  But it's better than nothing.