Move addDriveDialog() and friends to their own class
The reason behind this is to make these functions available for using
them in rescue mode too.

Note to reviewers, the code of addDriveDialog() and friends has not
changed a single letter it was only moved to the new textw/add_drive_text.py
file (and some boilerplate was added on the top to make it into a new class).