Add a script for running pylint on anaconda
This patch introduces a runpylint.sh script which invokes pylint in such
a way that it is usable with anaconda. Amongst other things it filters
out any false positives listed (regex) in the pylint-false-positives file.

If any problems where found the scripts prints them to stdout and
saves them in pylint-log. In this case it will also have a non 0
exit status (for future make archive integration).