Do not append "rhgb quiet" to s390 boot loader config (#570743)
Follow-on to 78d3854021245b1cb612be564d874fa8b2dca671.  As pointed out
on the mailing list, this problem was already discussed in another bug
and the correct fix was never actually pushed.  So here it is.  Put
parens around the ts.dbMatch tests and rely on the console on virtual
terminal test.

Related: rhbz#526364