lvm: check resizing against format's targetSize (#580171).
lv.format.currentSize remains on the original value all the time. This
means that when the user resizes the LV once, currentSize stays the same
and the test for change in lvm_dialog_gui.py always succeeds every time
the user reopens the LV edit dialog. This causes a traceback just below,
in ActionResizeFormat.