Check for fs utils when determining if an fs can be resized (#572431)
NTFS is not officially supported in RHEL, which means ntfsresize and
other tools are not available in the install media.  To determine if a
filesystem can be resized, we need to check three things:  is _resizable
set to True, does the filesystem exist, and are the filesystem utils
available.  The first two are checked by DeviceFormat.resizable(), make
sure utilsAvailable() is checked in FS.resizable().