Check for not having found any disks after populating the tree (#583906)
It is possible for us to not find any disks post the filter ui,
this can happen in 3 ways:
1. kickstart install (no filter ui, ignoreDisks)
2. text install (no filter ui)
3. gui install disk(s) selected in filter ui have no recognizable format
   and user clicked ignore

When this happens we currently backtrace in the clearpart ui as that
expects there to be atleast one disk. As there is very little we can
do without atleast one disk check for this condition in storageInitialize()
and allow the user to go back or exit the install. In the cmdline kickstart
install case this will result in the no disks found message being printed
and the installation stopping (it will trigger the cannot have a question in
cmdline mode code).