Enable networking in stage 2 using nm-c-e (#520146)
Key patch of series. enableNetwork method is the root of all other
changes in the patchset. The method is used both for network enabling
in stage 2 triggered by need of network (repos, network storage, etc..)
and [Configure Network] button

What we select in stage 2 network UI is:
- device to be activated (only in case of network enabling), this is
  controlled by ONBOOT ifcfg setting, note that devices actually
  activated depend on editing in nm-c-e
- devices to be controlled by NM (and thus visible in nm-c-e),
  this is set by NM_CONTROLLED ifcfg setting

A dialog for that is part of the patch, it needs some polishing.

bringUp method had to be split into write and waitForConnection
because in case of just config, without actually activating any
device in nm-c-e, we don't want to wait for activation.