Fetch ks files over NFS when ksdevice is not given (#541873)
Fixes up a problem that occurs when users pass:

    ks=nfs:HOST:PATH

But do not pass ksdevice.  The kickstartNetworkUp() function returns
early because NetworkManager is running and has a connection, but when
loader tries to pull the file over NFS, it looks for the network device
in loaderData->netDev, but kickstartNetworkUp() never filled that
information in.

Seems to only occur on s390x.  Tested on both s390x and x86_64.