Offer to format unformatted DASD devices (#560702)
On s390, if you had a single DASD that needed dasdfmt run, you would get
a ZeroDivisionError traceback because the self._totalCylinders value was
zero.  The problem was caused by the DASD.totalCylinders property and
how it initialized.  It would initialize itself on the first call, but
since we'd already fired off a dasdfmt process for the device, the one
running to capture the cylinder count couldn't get access to the device.

Users with more than one DASD needing a dasdfmt would not see the
traceback, but would have an incorrect total cylinder count, so 100%
would be reached when n-1 devices had been formatted.

Honor the exclusiveDisks list in the storage object and log the
/dev/disk/by-path alias for the device when logging the info message
about what device was found that was unformatted.

[Originally worked up in November 2009, I've updated the patch a bit to
work with our current code.  Tested in text mode and gui mode.  If you
have any unformatted DASD devices, anaconda prompts you and asks if you
want to format them.]