Bring up network for local hd vnc kickstart installs (#522064)
Where people come up with these use cases is beyond me.  Here we go:

1) Prepare an ext2/3/4 volume on the hard disk and drop the install DVD
   iso as well as images/install.img on there.
2) Drop a kickstart file on this volume called ks.cfg.  Make sure it:
       a) Has *no* 'network' lines
       b) Uses 'harddrive --partition=DEVICE --dir=/PATH'
       c) Contains a valid 'vnc' line
2) Copy vmlinuz and initrd.img to the /boot directory for the system.
3) Edit boot loader configuration to boot the installer kernel+ramdisk.
4) Reboot system, choose installer and pass:
       ks=hd:DEVICE:ks.cfg

Without this patch, the network is never configured during loader
because both the install media and kickstart file are provided locally.
However, we do need to bring up the network for vnc, so look for a vnc
line in the kickstart file and make sure the network is brought up.