Use separate method for copying network configuration to system (#520146).
Former Network.write was used to write config both to anaconda
environment and to installed system. It became quite messy and
unreadable doing both things as writing to installed system was just
copying the file from anaconda root to system root for all files
except ifcfg-DEVICE. Using two separate methods makes clear what
we are doing.

Also there is a question if we need to copy the network config
to system twice as we are doing - first in preinstall step,
and then in writeconfig step. The latter is not necessary I think
and seem to be actually a noop.