Use separate method for copying network configuration to system (#520146).
Former Network.write was used both to write config to anaconda
environment and to installed system. It became quite messy and
unreadable doing both things - writing to installed system was just
copying the file from anaconda root to system root for all files
except ifcfg-DEVICE.

This way, using two separate methods it is clear what we are doing.

Also there is a question if we need to copy the network config
to system twice as we are doing - first in preinstall step,
and then in writeconfig step. The latter is not necessary I think
and seem to be actually a noop (we are copying the file only
if it is not there yet).