Wait for specific activated network devices (#520146).
We used to activate just one device when enabling network,
so it was ok to use waitForConnection (wait just for any
connection). Now that more devices can be set to be activated
in nm-c-e (by checking [] Connect automatically), we should

(1) Wait for all devices activated by nm-c-e this way, otherwise
we can fail with traceback in next steps reading device
configuration.

(2) When enabling network make sure the device selected by user
in combobox was activated, not just any device.