Read in network settings correctly, as configured by linuxrc.s390
This patch fixes some old fallout from the move to NetworkManager.
Without this patch, loaderData never gets fully populated with the
network settings from linuxrc.s390.  Some backstory:

Before the move to NM, linuxrc.s390 wrote out /tmp/netinfo and loader
read that file it to populate loaderData.  By reading in these values,
loader would not have to prompt the user again for the same network
settings.

With the move to NM, this function was changed to read the only ifcfg
file in /etc/sysconfig/network-scripts and gather settings.  We cannot
always assume one file in /etc/sysconfig/network-scripts, so this patch
addresses that problem and makes some other improvements.

1) Read /etc/sysconfig/network-scripts/ifcfg-DEVICE where DEVICE is the
   value returned by getenv("DEVICE").
2) Remove old parsing code and use g_file_get_contents(), g_strsplit(),
   and g_shell_unquote().

Tested on s390x, which is the only platform that would call this
function, and it successfully elmininates the double prompting of
network settings.

Related: rhbz#522064