Remove partitions after unpartitioned non-partition devices. (#592337)
This basically removes a bunch of comparisons that would previously
return 0. The problem is that the 0 return value is often used as a
basis for further sorting, which can lead to incorrect ordering since
python's sort function does not compare items exhaustively.

Yes, I know. Topoligical sort is coming soon.