storage: LUKSDevice takes req_grow after its slave (#589450).
without the patch following ks line is a problem:
part pv.hd --grow --size=1 --encrypted --passphrase=firsttest

this creates a VG on a LUKSDevice that has a PartitionDevice as its
"slave" device. the partition device has 'req_grow' set to true as to True
but not the luks device so adding any logical volume to the VG fails with
insufficient space.