Advance line pointer & don't strdup(val) on error in readNetInfo
Fix two issues in the readNetInfo() change.  First, if the line lacks an
equal sign, increment the line pointer to the next line before
continuing.  Second, if g_shell_unquote() errors, we do not want to
strdup(val) on the basis of the field, just ignore the line entirely,
log it, and continue.

Related: rhbz#522064