Make Format and Resize checkboxes mutually exclusive (#589977)
In the Edit Partition dialog, disable Resize if Format is checked and
disable Format if Resize is checked.  Users cannot check both Format and
Resize and have it perform a useful operation.  Technically, it's
possible, but our UI workflow in anaconda does not really allow an easy
way to accomplish this.  For now, prevent users from selecting both
Format and Resize.  If we want to let people shrink a filesystem and
reformat it, we can get creative with that later.