Offer to ignore unformatted DASDs rather than forcing exit (#580456)
Similar to the fix for finding unformatted DASDs in rescue mode, rename
the 'Exit Installer' button to 'Ignore'.  Installation continues fine
when you ignore unformatted DASDs.  A little backstory, at the point in
time when I was working on this patch, I chose to force exit because
continuing with some unformatted DASDs did not work well with the
storage code at the time.  Seems to work now, so the choices users get
are Format and Ignore.

Since this is similar to what the rescue mode change does, I removed the
special "rescue" return value and just have it return 1.

Related: rhbz#582638