cleanup booty x86 flag.serial handling (#589773)
Bug 589773 is caused by booty's x86 flag.serial handling not setting
self.serial at all, when serial is present on the cmdline, but no
console= argument it present and the system is not ia64.

Looking closer at the code I also noticed that it will try to
access the options list at index 1, even though it is empty, when
console=ttyS0 is passed rather then console=ttyS0,...

All in all the code was rather messy, so this patch rewrites it to
better readable, hopefully obviously correct code.