Remove raid clone option and code (#589957)
Remove the raid clone option from the create-storage menu, and also
remove the underlying code. It is broken, and this has not been noticed
until very recently as no-one uses it. So as discussed on irc it is better
to just remove it, otherwise the next time we accidentally break it,
it will likely once more remain broken for a long time.