Determine if an mdmember is biosraid earlier (#586298)
Currently kickstart installs to Intel (mdraid using)  BIOS RAId sets
with "clearpart --all" do not work.

handleUdevDeviceFormat() calls shouldClear() to determine if format
specific handling needs to be done and shouldClear() depends on the
biosraid flag of the mdmember format to decide wether or not to
clear mdmember formats with --all (through the format.hidden attribute).

However the biosraid flag gets set by the format specific handling
which gets skipped in this case because it has not been set yet. This
patch fixes this by setting the biosraid property of the mdmember
format earlier by passing it in into its __init__ method, which
seems a much cleaner solution in general.