Add the best package for this arch to the optional package selector (#591653).
Before, we were just having ayum.searchNevra return a list of packages and grab
whatever happened to be first on that list.  Sometimes it ends up being the
x86-64 package, and sometimes it ends up being the i686 one instead.  To be
more deterministic, we should return the best package for the architecture
we're running on always.