Handle vgs with duplicate names (#591469)
With the changed udev rules properly setting the vg-uuid for pvs to the
pvs vg id, rather then getting the vg uuid by vg-name, and thus using
the same uuid for all pvs even if there are duplicate vgs, the uuid
check in LVMVolumeGroupDevice._addDevice triggers.

Rather then making this throw an error, track that there is a duplicate
vg and make complete always return false (even if there happens to be
the right number of pvs in the 2 vgs combined) and then let
devicetree._handleInconsistencies() further handle this.