Use progressbar instead of waitwindow for repo setup (#584996)
We can set callback for URLGrabber - that is download part
of repo setup so let's use pulse progress bar here. For sack
setup, we don't have usable callback, and also it is pretty
quick so let's keep using wait window for it.

I am not using normal (growing) progress bar for repo setup
because I would have to update it per downloaded repo metadata
file which would be needlessly fine-grained feedback for user.