booty: Don't create device.map entries for devices backing /
Currently booty tries to create device.map entries for / even when
there is a separate /boot. There is no reason to have device.map
entries when there is a separate /boot, and booty trips over when
trying to do this for complicated layered setups.

Removing the creating of device.map entries for devices backing /
worksaround this falling over. Note the second patch in this set
actually fixes the specific case of falling over happening in bug
593718, and both are intended for rhel6-branch to make booty more
robust wrt this.

Related: rhbz#593718