Improve handling of auto and unknown types in fstab. (#577260)
For entries of type "auto", we will use the scanned type we got
from blkid. We'll also write the actual type, not "auto", to the
new fstab if we write one.

Entries whose type does not resolve to any type we know about, but
whose scanned type does, will be ignored. The obvious example is
an entry of type "ntfs-3g" for which blkid reported "ntfs". We
don't know what to do with that and, frankly, don't care.