Make sure we still have an elf interpreter after unmounting fs (#598222)
When moving to stage2 we rename the initrd /lib to /lib_old and add
a symlink from /lib to /mnt/runtime/lib. When shutting down we unmount
filesystems and then after unmount try to run mdadm to wait for
any Intel BIOS RAID arrays to have their metadata updated that they
were cleanly shutdown to avoid a resync of the entire array on the next boot.

But at this point /lib is a dangling symlink, so no elf interpreter gets
found (nor libc) and the mdadm call we do fails, leaving the array unclean.