Only kill init for reboot/halt and then exit linuxrc.s390
Like init.c, linuxrc.s390 should exit immediately after having
sent the corresponding signal to the init process when called
as reboot or halt (or poweroff in the future).
While it might often not be noticed, if linuxrc continues after
sending the signal since the system shuts down anyway,
there are practical cases where the signalling does not work
and linuxrc would otherwise startup yet another instance of the
installer. Even though, linuxrc detects that it does not run
with PID==1, we should not bet on its "test" mode
for the cases where it serves as replacement for reboot/halt.

Related: rhbz#571370