Show by-path names for DASD and zFCP, WWID for mpath (#580507)
In the partition editor UI, we get device names for DASD as "dasda",
"dasdb", and so on.  For zFCP, we get "sda", "sdb", and so on. The by-path
identifiers are more useful to s390 users, so display those on this
screen if we have them, otherwise fall back on the device path.

For mpath devices, try to display the WWID if we have it, otherwise show
the device path.

(Based on patch from David Cantrell <dcantrell@redhat.com>)