Tell user when nothing can be upgraded (#592605)
When no upgradable partitions and at least one non-upgradable partition
are found display a dialog to the user warning them that it cannot be
upgraded. Let them continue to install or exit.

So, for example, if the user has only a Fedora 9 install they will be
told to upgrade to intermediate releases before upgrading to the
current release.

If the user has Fedora 11 and Fedora 9 they will not be shown the warning
and Fedora11 will be presented as a potential upgrade candidate.