Get netdev name without CONFIG_SYSFS_DEPRECATED_V2 in linuxrc.s390 (#596826)
Linuxrc.s390 used to read the symlink net:<netdev> to get the name
of the just configured network device of type lcs or ctc.
This symlink only exists if the kernel is compiled with
CONFIG_SYSFS_DEPRECATED_V2.
Since more recent kernel are built without that option, we read
the content of directory 'net' which is another directory having the
name of the network device.