Rebind hybrid lcs/ctc network devices to correct driver if necessary (#596826)
There are two network device types 3088/08 and 3088/1f which can be
either lcs or ctc. Therefore, both device drivers lcs and ctcm are
loaded and the one that happens to get loaded first binds such devices.
If the user specified the device to be handled by the other driver,
then rebind the subchannels to the specified driver.

Related: rhbz#557522
Related: rhbz#561814