In the filter UI, also ignore devices that do not report a size (#594803).
We need to throw these devices out of the list before testing whether to display
the filter UI or not.  Otherwise, these devices will make the filter UI appear
but will not be included in the overall count so it will appear that there's no
point in having the UI displayed.

With this fix, we can remove the later CCISS size check.

Based on a patch from notting.