Raise an error when an md dev is not in the tree after scanning all slaves
As the comment already indicates when this happens something is seriously
wrong. I've seen this code get triggered one or 2 times with BIOS RAID and
in both cases the cause was a bug elsewhere in anaconda. So rather then
the current "just wing it" solution, just raise an exception when this
happens.