Handle OPTIONS in ifcfg files transparently in loader (#595388)
Loader does not need to handle or understand layer2 and portno so
consolidate that into opaque options. This will prevent issues such as
bug 577005 or commit 9caaca4 or bug 468755
and should be also transparent to future extensions in linuxrc.

Correctly parse OPTIONS whose value includes equal signs because
the values are attribute value pairs. See also bug 597205 and
commit 8549a36.