Autopart PVs require enough space for a default-sized partition.
When identifying disks that will contain autopart PVs we are only
checking that each disk contain at least 100MB of contiguous free
space, but then we are allocating partitions with a base size of
500MB for the PVs. Hook both pieces to the same source to prevent
confusion.

Note that we can still get caught in many cases, eg: 900MB of free
space on the device /boot will be allocated from. We detect 900MB of
free space, so we mark it for a new PV. We allocate /boot, and then
don't have enough space left to allocate the PV.