Support for converged traffic during install to FCoE LUN (#604763)
With DCB it is possible to use one NIC for both FCoE and regular
IPv4 / IPv6 traffic. This patch allows selecting a NIC which is
already configured for IPv4 / IPv6 traffic for FCoE traffic.

Note that not all interactions with NetworkManager have been tested
yet, so converged traffic is considered a tech-preview feature for
RHEL-6.