Clean up proxy handling in yuminstall.py (#604137)
Make it possible to add proxy to AnacondaYumRepo objects from an
object with a proxy string and optionally username/password.

repos with no associated config file do not inherit the proxy from
the main yum.conf file, add the proxy setting passed from stage1
to AnacondaYumRepo objects that we create.

Resolves: rhbz#604137