Reset resolver after network device activation (#614001)
We need to reset resolver using our isys.resetResolver in rhel6
because curl is built without support for c-ares. If/when curl
supporting c-ares is pulled into rhel6, urlgrabber.reset_curl_obj()
call will be needed for resetting, so I also add the call to some
places where it is missing.