anaconda: Disable X server regenerations (#609245)
Otherwise, we lose arbitrary amounts of X server state, since we're not
waiting for metacity before carrying on; and any driver bugs that cause
crashes on regeneration will cause graphical install to fail entirely.

This varies slightly from the usual policy of doing what the installed
system would do, but I'm willing to turn a blind eye.  The more correct
fix would be waiting for one of the properties metacity creates to exist
on the root window.  This is harder than you think.

Signed-off-by: Adam Jackson <ajax@redhat.com>