Don't crash when putting mpath devices into the filter name cache (#597223).
The big problem here is that identifyMultipaths returns a list of lists of
components, whereas the name cache stuff expects just a regular list.
Therefore, I need to first flatten the component structure when creating
the cache and also check every component of an mpath device to see if it's
in the cache before deciding a device is new or not.