Don't skip method selection if using media with no packages (#611154).
Combined with removing "askmethod" from the boot arguments in scripts/, this
should make it so that booting with boot.iso or pxeboot will still prompt for
the installation method while all other ways of booting will continue as
intended.

By checking for stage2 on media this early, the url variable gets set.  If the
media also does not contain packages, method and skipMethodDialog will be unset.
Therefore STEP_METHOD will get run.  However, STEP_METHOD only sets method.
url remains set to whatever we found earlier.  So, STEP_STAGE2 will bail out
early.

The end result is that we use the stage2 from the boot media but refer to the
method given in the UI for the installation source.