Deal with media only for media repo package failures (#573492)
This should take care e.g. of the case when hd: or cdrom: repo
and additional non-media repo is used and download of package from
non-media repo fails which would lead to traceback.

The anacondaBaseURLs repo property is quite new in the repo code,
so now we can take advantage of having it there.