logging: autodetect the virtio-serial port.
Discussed with jlaska and amit. Removes the virtiolog= command line
parameter again, istead we will always look for
/dev/virtio-ports/org.fedoraproject.anaconda.log.0 and forward logs to it
if it exists. In theory, with some cooperation from the libvirt team, this
could ensure every F14 installation has remote logging set up, ready to be
picked on the host.

Related: rhbz#576439