Skip the askmethod dialog on boot.iso+method= (#611154).
The previous method selection patch had the unfortunate side effect of
causing the method selection screen to still appear for kickstart and
boot.iso installs when method=/repo= was provided.  This patch just tweaks
the conditionals a bit to correct that, plus add some debugging to make
this easier to debug in the future.